Remove NearestPathResolver and move its logic to Pimcore\Model\Document\Service
